utils/simplify_utils.py
from openai import AsyncOpenAI
import os
from dotenv import load_dotenv
from typing import List
import re

# Load environment variables
load_dotenv()

# Initialize OpenAI client
openai_client = AsyncOpenAI(api_key=os.getenv("OPENAI_API_KEY"))

async def process_text(text: str) -> str:
    """
    Process and simplify the input text using OpenAI's GPT-4.
    """
    try:
        response = await openai_client.chat.completions.create(
            model="gpt-4",
            messages=[
                {
                    "role": "system",
                    "content": "You are a text simplification assistant. Simplify the given text for clarity and brevity while maintaining the main ideas and key information. Make the text more accessible and easier to understand."
                },
                {
                    "role": "user",
                    "content": text
                }
            ],
            max_tokens=200,
            temperature=0.7
        )
        return response.choices[0].message.content.strip()
    except Exception as e:
        raise Exception(f"Error processing text with OpenAI: {str(e)}")

def get_word_complexity(word: str) -> float:
    """
    Calculate the complexity of a word.
    This is a placeholder function that can be enhanced with more sophisticated
    word complexity analysis.
    """
    # TODO: Implement word complexity analysis
    return 0.0

def get_simpler_synonyms(word: str) -> List[str]:
    """
    Get simpler synonyms for a word.
    This is a placeholder function that can be enhanced with a thesaurus API
    or word database.
    """
    # TODO: Implement synonym lookup
    return [] 