requirements.txt
fastapi==0.109.2
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pydantic[email]==2.6.1
sqlalchemy==1.4.23
pymysql==1.0.2
python-dotenv==0.19.0
autogen==0.1.0
openai==0.27.0
python-magic==0.4.27
qdrant-client==1.13.3 